database files moved to lib
